+++ app/src/test/java/com/pechenkin/travelmoney/test/TripForTest.java
@@ -104,7 +104,7 @@ public class TripForTest implements Trip {
 
 
     @Override
-    public void addCost(Member member, Member toMember, String comment, double sum, String image_dir, Date date, boolean isRepayment) {
+    public void addCost(Member member, Member toMember, String comment, int sum, String image_dir, Date date, boolean isRepayment) {
 
     }
 }
